refactor: move tool
